+++ R/class.R
@@ -11,7 +11,11 @@
 #'   * The R7 class, like [R7_object].
 #'   * An S3 class wrapped by [new_S3_class()].
 #'   * A base type, like `logical`, `double`, or `character`.
-#'
+#' @param package Package name. It is good practice to set the package
+#'   name when exporting an R7 class from a package because it includes
+#'   the package name in the class name when it's used for dispatch. This
+#'   allows different packages to use the same name to refer to different
+#'   classes.
 #' @param constructor The constructor function. Advanced use only.
 #'
 #'   A custom constructor should call `new_object()` to create the R7 object.
@@ -84,6 +88,7 @@
 new_class <- function(
     name,
     parent = R7_object,
+    package = NULL,
     properties = list(),
     constructor = NULL,
     validator = NULL) {
@@ -97,6 +102,10 @@ new_class <- function(
          "`parent` must be an R7 class, S3 class, or base type, not %s.", class_friendly(parent)),
        call. = FALSE
      )
+  }
+
+  if (!is.null(package)) {
+    check_name(package)
   }
 
   if (!is.null(constructor) && !is.null(parent)) {
@@ -119,6 +128,7 @@ new_class <- function(
   # Must synchronise with prop_names
   attr(object, "name") <- name
   attr(object, "parent") <- parent
+  attr(object, "package") <- package
   attr(object, "properties") <- all_props
   attr(object, "constructor") <- constructor
   attr(object, "validator") <- validator
@@ -126,6 +136,11 @@ new_class <- function(
 
   global_variables(names(all_props))
   object
+}
+globalVariables(c("name", "parent", "package", "properties", "constructor", "validator"))
+
+R7_class_name <- function(x) {
+  paste(c(x@package, x@name), collapse = "::")
 }
 
 check_R7_constructor <- function(constructor) {